backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aniso8601==7.0.0
    # via graphene
asgiref==3.4.1
    # via django
django==3.2.6
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   graphene-django
django-cors-headers==3.7.0
    # via -r requirements.in
django-environ==0.4.5
    # via -r requirements.in
graphene==2.1.9
    # via graphene-django
graphene-django==2.15.0
    # via -r requirements.in
graphql-core==2.3.2
    # via
    #   graphene
    #   graphene-django
    #   graphql-relay
graphql-relay==2.0.1
    # via graphene
promise==2.3
    # via
    #   graphene-django
    #   graphql-core
    #   graphql-relay
psycopg2==2.9.1
    # via -r requirements.in
pytz==2021.1
    # via django
rx==1.6.1
    # via graphql-core
singledispatch==3.6.2
    # via graphene-django
six==1.16.0
    # via
    #   graphene
    #   graphene-django
    #   graphql-core
    #   graphql-relay
    #   promise
    #   singledispatch
sqlparse==0.4.1
    # via django
text-unidecode==1.3
    # via graphene-django
